pyproject.toml
[project]
name = "infobus-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.5.3",
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=5.2.6",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.13.0",
    "watchdog>=6.0.0",
]
